wip refactoring

+++ frontend/js/panier.js
@@ -1,16 +1,18 @@
+// window.addEventListener("DOMContentLoaded", fn)
+
+// Split code into functions that have own responsabilities
+// Avoid side effects into same function
+
 //  --------------------Données des produits enregistrer dans le local storage-------------------------//
 let saveProductOnLocalStorage = JSON.parse(localStorage.getItem("teddyBear"));
 
-
 // --------------------AFfichage produit du panier-------------------------//
 const showEmptyBasket = document.querySelector("#container_list_product");
-const showFullBasket = document.querySelector("#show_basket_proudct")
-
+const showFullBasket = document.querySelector("#show_basket_proudct");
 
 // Si le panier est vide
-if(saveProductOnLocalStorage === null){
-  let emptyBasket = 
-  `
+if (saveProductOnLocalStorage === null) {
+  let emptyBasket = `
   <table align = center>
    <thead>
     <tr>
@@ -18,16 +20,15 @@ if(saveProductOnLocalStorage === null){
     </tr>
    </thead>
   </table>
-  `
+  `;
   showEmptyBasket.innerHTML = emptyBasket;
 }
 //Affichage des produits choisis dans un tableau
-else{
+else {
   let fullBasket = [];
- 
-  for(i = 0; i < saveProductOnLocalStorage.length; i++){
-    fullBasket += 
-`
+
+  for (i = 0; i < saveProductOnLocalStorage.length; i++) {
+    fullBasket += `
    <tr>
     <td class="product_basket">${saveProductOnLocalStorage[i].productName}</td>
       <td class = "option_basket">${saveProductOnLocalStorage[i].productOption}</td>
@@ -36,13 +37,12 @@ else{
             <td><button class = "delete_btn" >Supprimer</button> </td>
    <tr>
 `;
-}
+  }
 
-if(i == saveProductOnLocalStorage.length){
-  showFullBasket.innerHTML = fullBasket;
+  if (i === saveProductOnLocalStorage.length) {
+    showFullBasket.innerHTML = fullBasket;
+  }
 }
-}
-
 
 //  --------------------Création du bouton SUPPRIMER L ARTICLE-------------------------//
 // let deleteProduct = document.querySelectorAll(".delete_btn");
@@ -55,7 +55,6 @@ if(i == saveProductOnLocalStorage.length){
 //   })
 // }
 
-
 //  --------------------Création du bouton SUPPRIMER LE PANIER-------------------------//
 let deleteBasket = `   
 <tr>
@@ -63,30 +62,35 @@ let deleteBasket = `
 </tr>
 `;
 
-showFullBasket.insertAdjacentHTML("beforeend" , deleteBasket);
+showFullBasket.insertAdjacentHTML("beforeend", deleteBasket);
 
-let deleteAll = document.querySelector(".deleteAll")
-
-deleteAll.addEventListener('click' , (e)=>{
-  e.preventDefault;
-  e.stopPropagation;
-  localStorage.removeItem("teddyBear");
-  alert('Vous avez supprimer votre panier');
-  window.location.href = "panier.html";
-})
+let deleteAll = document.querySelector(".deleteAll");
+if (deleteAll) {
+  deleteAll.addEventListener("click", (e) => {
+    e.preventDefault;
+    e.stopPropagation;
+    localStorage.removeItem("teddyBear");
+    alert("Vous avez supprimer votre panier");
+    window.location.href = "panier.html";
+  });
+}
 
 //  --------------------Calcul du prix total du panier-------------------------//
-let totalBasketPrice = [];
+let totalBasketPrice = [1, 2, 3, 4];
 
-for(i = 0; i < saveProductOnLocalStorage.length; i++){
-
+if (saveProductOnLocalStorage) {
+  for (i = 0; i < saveProductOnLocalStorage.length; i++) {
     let totalProductPrice = saveProductOnLocalStorage[i].price;
 
-    totalBasketPrice.push(totalProductPrice)
-    };
+    totalBasketPrice.push(totalProductPrice);
+  }
+}
 
-const reducer = (accumulator , currentValue) => accumulator + currentValue;
-const totalPrice = totalBasketPrice.reduce(reducer);
+function summarize(acc, cur) {
+  return Number.parseInt(acc) + Number.parseInt(cur);
+}
+
+const totalPrice = totalBasketPrice.reduce(summarize, 0);
 console.log(totalPrice);
 
 let basketPrice = `   
@@ -95,7 +99,4 @@ let basketPrice = `
 </tr>
 `;
 
-showFullBasket.insertAdjacentHTML("beforeend" , basketPrice);
-    
-
-
+showFullBasket.insertAdjacentHTML("beforeend", basketPrice);
